refactor: moved user search form component to shared folder
